admin-API: add route and handler for listing KMS master keys
This commit adds an admin API route and handler for
listing objects encrypted with master keys resp.
master keys of objects.

Therefore, the client specifies the KMS key ID (when
empty / not set the server takes the currently configured
default key-ID) and a bucket/object prefix and the server
tries to list all objects encrypted with that key. Therefore,
the server returns a (new-line separated) list of JSON encoded
values.