Decouple ServiceUpdate to ServerUpdate to be more native
The change now is to ensure that we take custom URL as
well for updating the deployment, this is required for
hotfix deliveries for certain deployments - other than
the community release.

This commit changes the previous work d65a2c672518a8af5f37f07c3abf18222df1839a
with newer set of requirements.

Also deprecates PeerUptime()