add KeyStore interface and Vault env. variable
This commit adds the `KeyStore` interface
to load and store secret keys at an external
key-value store.

Further, it adds the `MINIO_SSE_VAULT_KEY_STORE`
env. var. to specify the path to the key store.

In general a MinIO cluster can use either a KMS
(with master keys) or a Key Store for key management - but
not both at the same time.

Follow up on #8259

Co-Authored-By: Krishnan Parthasarathi <krisis@users.noreply.github.com>