WIP: add basic support for a secret KeyStore
This commit adds basic support for fetching
and storing secret keys (for SSE-S3) at an
external key store.

Whenever a object is deleted we should try to
delete the secret key from the key store.
The current approach follows an eventual consistency
model where we delete the secret key after we have
deleted the object from the backend successfully.
However, it may be better to only delete the object
from the backend *if* we have deleted the secret key
from the key store.