KMS Admin-API add route and handler for KMS key rotation
This commit adds an admin API route and handler for
rotating master master keys of objects.

Therefore, the client specifies the old and new KMS key ID
and a bucket/object prefix and the server tries to re-encrypt
the object encryption keys of all objects encrypted with the
old master key ID. The server returns a (new-line separated)
list of JSON encoded values - each representing a successful
key rotation.