ecc: add new package and remove io.Pipe in object-API
This commit adds the new `ecc` (error correction code) package
that implements primitives for MinIO's erasure coding and bitrot
protection. The main purpose of `ecc` is to provide a simpler
and well-documented implementation of erasure-coding and bitrot
protection primitives to reduce the overall code complexity.

This commit does not remove any previous code since it is still
used by e.g. healing. However, the plan is to move / re-implement
encoding and healing to/in `ecc`in the future.

The `ecc`package documentation can be viewed via:
  1. `godoc -http=":8080" -goroot=<your-go-root-path>`
  2. http://localhost:8080/pkg/github.com/minio/minio/cmd/ecc