posix: unify ReadFile storage API
This commit unifies the ReadFile storage API
such that "legacy" and streaming bitrot uses
the same `ReadFile` API.

This is a refactoring that makes future changes
to bitrot verification and erasure coding easier.
In particular, it simplifies moving streaming
bitrot verification to the "server side" / posix
layer.

With the `ecc` package this commit tries to organize
and simplify the structure of the erasure coding and
content verification layers in the MinIO server.