usage: Fix buckets count calculation when no object is present
XL crawling wrongly returns a zero buckets count when there is
no objects uploaded in the server yet. The reason is data
of the crawler of posix returns invalid result when all disks has zero objects.

A simple fix is to always pick the crawling result of the first disk
but choose over the result of the disk which has the most objects in it.