Time getSize and use to estimate latency
Remove the random sleep. This is running in 4 goroutines, so mostly doing nothing.

We use the getSize latency to estimate system load, meaning when there is little load on the system and we get the result fast we sleep a little.

If it took a long time we have high load and release ourselves longer.

We are sleeping inside the mutex so this affects all goroutines doing IO.