fix: Avoid crash when there is an error testing a target notif
RegisterNotificationTargets() cleans up all connections that it mades to
notification targets when an error occurs during its execution.

However there is a typo in the code that makes the function to always
try to access to a nil pointer in the defer code since the function
in question will always return nil in the case of any error.

This commit fixes the typo in the code.