introduce WalkOptions to customize object traversal
This commit adds a `WalkOptions` struct to allow
customization of object traversal via the `Walk`
method.

This commit only adds the recursive flag. More options
may be added in the future.

Further, this commit changes who has to create the
`ObjectInfo` channel used to communicate the object
traversal results. Now, the caller of the `Walk` just
obtains a receive-only channel and can iterate over it.
The `Walk` implementation itself is now responsible to
create the channel.