create the final file with mode 0666 for multipart-uploads
MinIO Gateway for NAS creates non-multipart-uploads with mode 0666.
In cmd/fs-v1-helpers.go:324 mode 0666 is used to create new files.
But multipart-uploads are created with a differing mode of 0644.

Both modes should be equal! Else it leads to files with different
permissions based on its file-size. This patch solves that by
using 0666 for both cases.