Use default keepalive check interval
I suspect there was some misunderstanding with the Dialer KeepAlive value:

```
	// KeepAlive specifies the interval between keep-alive
	// probes for an active network connection.
```

So this value specifies how often to check if a connection is still alive.

Since idle connections were killed after 1 minute this didn't make sense for most callers.
We remove it from parameters and set it to the default 15 seconds.

The idle timeout of 1 minute is a bit low, so it gets bumped to 5 minutes.