Merge nested hash readers
The `ioutil.NopCloser(reader)` was hiding nested hash readers.

We make it an `io.Closer` so it can be attached without wrapping and allows for nesting, by merging the requests.

Not really sure what the `strictCompat` is actually about, since nobody documented it, so just keeping it as is and hoping it does what it is supposed to.

Added tests.