Group healing new disks features in one place
Reformatting disks, healing .minio.sys objects.. and recreating
buckets (e.g. after a bucket expansion) is all moved in the
background heal of new disks code.

The functional change in this is that healing will only happen
when *all* nodes are online.

Another function change is that there will be one leader node
which will coordinate this (healing new disks).