Improve listobjects performance
* Only fetch objects from a limited number of disks.

For now it is set to max 3 disks. We should at least be able to reduce it to len(disks)/2.

* Only Stat objects we are interested in.

Filter out unwanted prefixed early, start at marker, listen for cancellations.

We keep upstream filtering so we don't have to upgrade REST version.

* List directories concurrently

When querying multiple disks in XL mode do it concurrently.

No benchs yet.