Cache disk info
Add a 1 second cache for disk info.

This value is requested on every upload when there are multiple zones.

Since this will result in an RPC call to every remote disk this scales quite badly in a distributed setup.

2 servers, localhost only. In large distributed setups a much bigger gains can be expected.

```
Operations: 21743 -> 22454
* Average: +3.28% (+0.0 MiB/s) throughput, +3.28% (+11.9) obj/s
* Fastest: +3.37% (+0.0 MiB/s) throughput, +3.37% (+13.0) obj/s
* 50% Median: +3.03% (+0.0 MiB/s) throughput, +3.03% (+11.2) obj/s
* Slowest: +8.03% (+0.0 MiB/s) throughput, +8.03% (+22.8) obj/s
```

For easy management of this a generic helper has been added.