Reconnect rest client async
Currently marking a client as offline has no other effect than all calls trying to reconnect anyway, so it is a waste of resources.

When encountering a rest network error we spawn a goroutine that will attempt to reconnect every second and reject requests until the server is reconnected.

Closing a client will also keep the connection closed whereas now it has no effect.

The effect should be that an offline server will make the rest of the cluster behave much better since requests will fail instantly and not put additional pressure on the system.