Support for context with long running Disk calls
ListObjects, Walkers do not timeout quickly if the
client has disconnected, this leads to quite a lot
of resources being spent on the drives perpetually
until some form to timeout happens. This can be
quite slow.

Many such operations can lead to quickly overwhelming
the server with too many List requests which are
yet to be finished.

To avoid this pass down context everywhere, and timeout
as connections timeout or reach deadline whichever comes
first.