Do not select zones with <5% free after upload
Looking into full disk errors on zoned setup. We don't take the 5% space requirement into account when selecting a zone. That is easy to fix.

The interesting part is that even considering this we don't know the size of the object the user wants to upload when they do multipart uploads.

It seems quite defensive to always upload multiparts to the zone where there is the most space since all load will be directed to a part of the cluster.

In these cases we make sure it can at least hold a 100MiB file and we disadvantage fuller zones more by subtracting the expected size before weighing.