heal: Redirect heal request based on client token
When a manual healing is triggered, one node in a cluster will take care
of that. mc regularly sends new admin requests to fetch the status of
the going healing process, but a load balancer could land the heal
request to a node that is not doing the healing request and not aware
about it.

This PR will redirect a request to the node based on the node index
found in the client token itself.