Copy Object: Allow multipart uploads
The `minio-go.Core` does not switch automatically to multipart uploads as the `minio-go.Client` does.

This means that no matter the size files would always be attempted to be uploaded with a single PUT calls.

We set last modified to now since this is not returned. ETag is overwritten right away anyway.

Cherrypick for release.