Allow Content-Md5 to be extracted as metadata
The correct behaviour of e.g. Azure gateway depends on this.

Some applications rely on Content-Md5 field being preserved, but
for objects greater than azureBlockSize/2 it is not being filled
on the Azure side, leading to ETag field being used instead, which
doesn’t match the value those applications expect.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>