Add storage layer contexts
Add context to all (non-trivial) calls to the storage layer.

Contexts are propagated through REST client.

`context.TODO()` is left in place for the places where it needs to be added to the caller.

`endWalkCh` could probably be removed from the walkers, but no changes so far.

The "dangerous" part is that now a caller disconnecting *will* propagate down, so a "delete" operation will now be interrupted. In some cases we might want to disconnect this functionality so the operation completes if it has started, leaving the system in a cleaner state.

 In a few places this is disconnected for cleanup operations.