use trailers for locker communication
We use trailers here to ensure that
client simply waits for the last trailer
for remote locker to be successful.

Client needs to see "Trailer" value
to be "Success" to consider locking
a successful operation, this change
is to ensure that there is no room for
races to occur when client actually
gives up upon timeout but server ends
up locking successfully anyways.