Fix write quorum calculation for non-objects writing
When the number of disks is odd, the calculation of data/parity disks
for buckets creation become incorrect.

For example, if we have 5 disks, and EC = 2, then it is enough to
require 3 disks to create a bucket but the current code wants 4
disks due to its calcuation.

This commit fixes that behavior.