List V1 requests unnecessarily fail with offline nodes
List V1 requests are actually redirected to a specific node, depending
on the bucket name. The purpose of this behavior is to optimize long
listing.

However, the current code sends a Bad Gateway error if the
target node is offline, which is a bad behavior because it means
that the list request will fail, although this is unecessary since
we can still use the current node to list as well (the default behavior
 without using proxing optimization)

Currently you can see mint fails when there is one offline node, after
this PR, mint will always suceeed.