Ignore config values from unknown subsystems
When encountering a config value from an unknown subsystem the servers will crash with `panic: assignment to nil map`.

Instead log a message and ignore the config.

While this is mostly a thing that will happen on downgrades it the server shouldn't crash when an unknown config subsystem is seen.